Struct vortex::array::TemporalArray

pub struct TemporalArray { /* private fields */ }

An array wrapper for primitive values that have an associated temporal meaning.

This is a wrapper around ExtensionArrays containing numeric types, each of which corresponds to either a timestamp or julian date (both referenced to UNIX epoch), OR a time since midnight.

Arrow compatibility

TemporalArray can be created from Arrow arrays containing the following datatypes:

Time32

Time64

Timestamp

Date32

Date64

Anything that can be constructed and held in a TemporalArray can also be zero-copy converted back to the relevant Arrow datatype.

Implementations

impl TemporalArray

pub fn new_date(array: Array, time_unit: TimeUnit) -> Self

Create a new TemporalArray holding either i32 day offsets, or i64 millisecond offsets that are evenly divisible by the number of 86,400,000.

This is equivalent to the data described by either of the Date32 or Date64 data types from Arrow.

Panics

If the time unit is milliseconds, and the array is not of primitive I64 type, it panics.

If the time unit is days, and the array is not of primitive I32 type, it panics.

If any other time unit is provided, it panics.

pub fn new_time(array: Array, time_unit: TimeUnit) -> Self

Create a new TemporalArray holding one of the following values:

i32 values representing seconds since midnight

i32 values representing milliseconds since midnight

i64 values representing microseconds since midnight

i64 values representing nanoseconds since midnight

Note, this is equivalent to the set of values represented by the Time32 or Time64 types from Arrow.

Panics

If the time unit is seconds, and the array is not of primitive I32 type, it panics.

If the time unit is milliseconds, and the array is not of primitive I32 type, it panics.

If the time unit is microseconds, and the array is not of primitive I64 type, it panics.

If the time unit is nanoseconds, and the array is not of primitive I64 type, it panics.

pub fn new_timestamp( array: Array, time_unit: TimeUnit, time_zone: Option<String>, ) -> Self

Create a new TemporalArray holding Arrow spec compliant Timestamp data, with an optional timezone.

Panics

If array does not hold Primitive i64 data, the function will panic.

If the time_unit is days, the function will panic.

impl TemporalArray

pub fn temporal_values(&self) -> Array

Access the underlying temporal values in the underlying ExtensionArray storage.

These values are to be interpreted based on the time unit and optional time-zone stored in the TemporalMetadata.

pub fn temporal_metadata(&self) -> &TemporalMetadata

Retrieve the temporal metadata.

The metadata is used to provide semantic meaning to the temporal values Array, for example to understand the granularity of the samples and if they have an associated timezone.

pub fn ext_dtype(&self) -> &ExtDType

Retrieve the extension DType associated with the underlying array.