Struct vortex::array::VarBinViewArray

pub struct VarBinViewArray { /* private fields */ }

Implementations

impl VarBinViewArray

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

impl VarBinViewArray

pub fn try_new( views: Array, buffers: Vec<Array>, dtype: DType, validity: Validity, ) -> VortexResult<Self>

pub fn buffer_count(&self) -> usize

Number of raw string data buffers held by this array.

pub fn view_slice(&self) -> &[BinaryView]

Access to the underlying views child array as a slice of BinaryView structures.

This is useful for iteration over the values, as well as for applying filters that may only require hitting the prefixes or inline strings.

pub fn view_at(&self, index: usize) -> BinaryView

pub fn views(&self) -> Array

Access to the primitive views array.

Variable-sized binary view arrays contain a “view” child array, with 16-byte entries that contain either a pointer into one of the array’s owned buffers OR an inlined copy of the string (if the string has 12 bytes or fewer).

pub fn buffer(&self, idx: usize) -> Array

Access one of the backing data buffers.

Panics

This method panics if the provided index is out of bounds for the set of buffers provided at construction time.

pub fn buffers(&self) -> Buffers<'_>

Retrieve an iterator over the raw data buffers. These are the BYTE buffers that make up the array’s contents.

Example

use vortex::array::VarBinViewArray;
let array = VarBinViewArray::from_iter_str(["a", "b", "c"]);
array.buffers().for_each(|block| {
    // Do something with the `block`
});

pub fn validity(&self) -> Validity

pub fn from_iter<T: AsRef<[u8]>, I: IntoIterator<Item = Option<T>>>( iter: I, dtype: DType, ) -> Self

Accumulate an iterable set of values into our type here.

pub fn from_iter_str<T: AsRef<str>, I: IntoIterator<Item = T>>(iter: I) -> Self

pub fn from_iter_nullable_str<T: AsRef<str>, I: IntoIterator<Item = Option<T>>>( iter: I, ) -> Self

pub fn from_iter_bin<T: AsRef<[u8]>, I: IntoIterator<Item = T>>(iter: I) -> Self

pub fn from_iter_nullable_bin<T: AsRef<[u8]>, I: IntoIterator<Item = Option<T>>>( iter: I, ) -> Self

pub fn bytes_at(&self, index: usize) -> VortexResult<Vec<u8>>