Module vortex::compute

Compute kernels on top of Vortex Arrays.

We aim to provide a basic set of compute kernels that can be used to efficiently index, slice, and filter Vortex Arrays in their encoded forms.

Every array variant has the ability to implement their own efficient implementations of these operators, else we will decode, and perform the equivalent operator from Arrow.

Modules

unary

Enums

Operator

SearchResult

Result of performing search_sorted on an Array

SearchSortedSide

Traits

AndFn

ArrayCompute

Trait providing compute functions on top of Vortex arrays.

CompareFn

FilterFn

IndexOrd

Len

MaybeCompareFn

OrFn

SearchSorted

SearchSortedFn

Searches for value assuming the array is sorted.

SliceFn

Limit array to start…stop range

TakeFn

Functions

and

compare

filter

Return a new array by applying a boolean predicate to select items from a base Array.

or

scalar_cmp

search_sorted

search_sorted_many

Search for many elements in the array.

search_sorted_u64

search_sorted_u64_many

slice

Return a zero-copy slice of an array, between start (inclusive) and end (exclusive).

take