Function vortex::compute::slice

pub fn slice(
    array: impl AsRef<Array>,
    start: usize,
    stop: usize,
) -> VortexResult<Array>

Return a zero-copy slice of an array, between start (inclusive) and end (exclusive).

Errors

Slicing returns an error if you attempt to slice a range that exceeds the bounds of the underlying array.

Slicing returns an error if the underlying codec’s slice implementation returns an error.