Trait vortex::compute::ArrayCompute

pub trait ArrayCompute {
    // Provided methods
    fn cast(&self) -> Option<&dyn CastFn> { ... }
    fn compare(
        &self,
        _other: &Array,
        _operator: Operator,
    ) -> Option<VortexResult<Array>> { ... }
    fn fill_forward(&self) -> Option<&dyn FillForwardFn> { ... }
    fn filter(&self) -> Option<&dyn FilterFn> { ... }
    fn scalar_at(&self) -> Option<&dyn ScalarAtFn> { ... }
    fn subtract_scalar(&self) -> Option<&dyn SubtractScalarFn> { ... }
    fn search_sorted(&self) -> Option<&dyn SearchSortedFn> { ... }
    fn slice(&self) -> Option<&dyn SliceFn> { ... }
    fn take(&self) -> Option<&dyn TakeFn> { ... }
    fn and(&self) -> Option<&dyn AndFn> { ... }
    fn or(&self) -> Option<&dyn OrFn> { ... }
}

Trait providing compute functions on top of Vortex arrays.

Provided Methods

fn cast(&self) -> Option<&dyn CastFn>

Implemented for arrays that can be casted to different types.

See: CastFn.

fn compare( &self, _other: &Array, _operator: Operator, ) -> Option<VortexResult<Array>>

Binary operator implementation for arrays against other arrays.

See: CompareFn.

fn fill_forward(&self) -> Option<&dyn FillForwardFn>

Array function that returns new arrays a non-null value is repeated across runs of nulls.

See: FillForwardFn.

fn filter(&self) -> Option<&dyn FilterFn>

Filtering function on arrays of predicates.

See: FilterFn.

fn scalar_at(&self) -> Option<&dyn ScalarAtFn>

Single item indexing on Vortex arrays.

See: ScalarAtFn.

fn subtract_scalar(&self) -> Option<&dyn SubtractScalarFn>

Broadcast subtraction of scalar from Vortex array.

See: SubtractScalarFn.

fn search_sorted(&self) -> Option<&dyn SearchSortedFn>

Perform a search over an ordered array.

See: SearchSortedFn.

fn slice(&self) -> Option<&dyn SliceFn>

Perform zero-copy slicing of an array.

See: SliceFn.

fn take(&self) -> Option<&dyn TakeFn>

Take a set of indices from an array. This often forces allocations and decoding of the receiver.

See: TakeFn.

fn and(&self) -> Option<&dyn AndFn>

Perform a boolean AND operation over two arrays

See: AndFn.

fn or(&self) -> Option<&dyn OrFn>

Perform a boolean OR operation over two arrays

See: OrFn.