Trait vortex::compute::SliceFn

pub trait SliceFn {
    // Required method
    fn slice(&self, start: usize, stop: usize) -> VortexResult<Array>;
}

Limit array to start…stop range

Required Methods

fn slice(&self, start: usize, stop: usize) -> VortexResult<Array>

Return a zero-copy slice of an array, between start (inclusive) and end (exclusive). If start >= stop, returns an empty array of the same type as self. Assumes that start or stop are out of bounds, may panic otherwise.