Module vortex::compute::unary

Traits

CastFn

FillForwardFn

Trait for filling forward on an array, i.e., replacing nulls with the last non-null value.

ScalarAtFn

SubtractScalarFn

Functions

fill_forward

scalar_at

scalar_at_unchecked

Returns a Scalar value without checking for validity or array bounds. Might panic OR return an invalid value if used incorrectly.

subtract_scalar

try_cast

Attempt to cast an array to a desired DType.