Trait vortex::compute::unary::FillForwardFn

pub trait FillForwardFn {
    // Required method
    fn fill_forward(&self) -> VortexResult<Array>;
}

Trait for filling forward on an array, i.e., replacing nulls with the last non-null value.

If the array is non-nullable, it is returned as-is. If the array is entirely nulls, the fill forward operation returns an array of the same length, filled with the default value of the array’s type. The DType of the returned array is the same as the input array; the Validity of the returned array is always either NonNullable or AllValid.

Required Methods

fn fill_forward(&self) -> VortexResult<Array>