Enum vortex::Array

pub enum Array {
    Data(ArrayData),
    View(ArrayView),
}

A central type for all Vortex arrays, which are known length sequences of typed and possibly compressed data.

This is the main entrypoint for working with in-memory Vortex data, and dispatches work over the underlying encoding or memory representations.

Variants

Data(ArrayData)

Owned Array with serialized metadata, backed by heap-allocated memory.

View(ArrayView)

Zero-copy view over flatbuffer-encoded Array data, created without eager serialization.

Implementations

impl Array

pub fn as_primitive(&self) -> PrimitiveArray

impl Array

pub fn tree_display(&self) -> TreeDisplayWrapper<'_>

impl Array

pub fn encoding(&self) -> EncodingRef

pub fn len(&self) -> usize

Returns the number of logical elements in the array.

pub fn is_empty(&self) -> bool

pub fn nbytes(&self) -> usize

Total size of the array in bytes, including all children and buffers.

pub fn child<'a>( &'a self, idx: usize, dtype: &'a DType, len: usize, ) -> VortexResult<Self>

pub fn children(&self) -> Vec<Array>

Returns a Vec of Arrays with all of the array’s child arrays.

pub fn nchildren(&self) -> usize

Returns the number of child arrays

pub fn depth_first_traversal(&self) -> ArrayChildrenIterator

pub fn cumulative_nbuffers(&self) -> usize

Count the number of cumulative buffers encoded by self.

pub fn all_buffer_offsets(&self, alignment: usize) -> Vec<u64>

Return the buffer offsets and the total length of all buffers, assuming the given alignment. This includes all child buffers.

pub fn buffer(&self) -> Option<&Buffer>

pub fn into_buffer(self) -> Option<Buffer>

pub fn into_array_iterator(self) -> impl ArrayIterator

pub fn into_array_stream(self) -> impl ArrayStream

pub fn is_encoding(&self, id: EncodingId) -> bool

Checks whether array is of a given encoding.

pub fn with_dyn<R, F>(&self, f: F) -> R
where F: FnMut(&dyn ArrayTrait) -> R,