Module vortex::iter

Structs

ArrayIteratorAdapter

Batch

FlattenedBatch

VectorizedArrayIter

Iterate over batches of compressed arrays, should help with writing vectorized code.

Constants

ITER_BATCH_SIZE

Traits

Accessor

Define the basic behavior required for batched iterators

ArrayIterator

A stream of array chunks along with a DType. Analogous to Arrow’s RecordBatchReader.

ArrayIteratorExt

Type Aliases

AccessorRef