Trait vortex::ArrayMetadata

pub trait ArrayMetadata:
    'static
    + Send
    + Sync
    + Debug
    + TrySerializeArrayMetadata
    + Display {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn as_any_arc(self: Arc<Self>) -> Arc<dyn Any + Send + Sync>;
}

Dynamic trait used to represent opaque owned Array metadata

Note that this allows us to restrict the (’static + Send + Sync) requirement to just the metadata trait, and not the entire array trait. We require ’static so that we can downcast use the Any trait. TODO(ngates): add Display

Required Methods

fn as_any(&self) -> &dyn Any

fn as_any_arc(self: Arc<Self>) -> Arc<dyn Any + Send + Sync>