Trait vortex::IntoArrayVariant

pub trait IntoArrayVariant {
    // Required methods
    fn into_null(self) -> VortexResult<NullArray>;
    fn into_bool(self) -> VortexResult<BoolArray>;
    fn into_primitive(self) -> VortexResult<PrimitiveArray>;
    fn into_struct(self) -> VortexResult<StructArray>;
    fn into_varbinview(self) -> VortexResult<VarBinViewArray>;
    fn into_extension(self) -> VortexResult<ExtensionArray>;
}

Trait for types that can be converted from an owned type into an owned array variant.

Canonicalization

This trait has a blanket implementation for all types implementing IntoCanonical.

Required Methods

fn into_null(self) -> VortexResult<NullArray>

fn into_bool(self) -> VortexResult<BoolArray>

fn into_primitive(self) -> VortexResult<PrimitiveArray>

fn into_struct(self) -> VortexResult<StructArray>

fn into_varbinview(self) -> VortexResult<VarBinViewArray>

fn into_extension(self) -> VortexResult<ExtensionArray>