Trait vortex::IntoCanonical

pub trait IntoCanonical {
    // Required method
    fn into_canonical(self) -> VortexResult<Canonical>;
}

Support trait for transmuting an array into its vortex_dtype::DType’s canonical encoding.

This conversion ensures that the array’s encoding matches one of the builtin canonical encodings, each of which has a corresponding Canonical variant.

Invariants

The DType of the array will be unchanged by canonicalization.

Required Methods

fn into_canonical(self) -> VortexResult<Canonical>