Trait vortex::variants::BoolArrayTrait

pub trait BoolArrayTrait: ArrayTrait {
    // Required methods
    fn maybe_null_indices_iter<'a>(
        &'a self,
    ) -> Box<dyn Iterator<Item = usize> + 'a>;
    fn maybe_null_slices_iter<'a>(
        &'a self,
    ) -> Box<dyn Iterator<Item = (usize, usize)> + 'a>;

    // Provided methods
    fn invert(&self) -> VortexResult<Array>
       where Self: Clone { ... }
    fn true_count(&self) -> usize { ... }
}

Required Methods

fn maybe_null_indices_iter<'a>(&'a self) -> Box<dyn Iterator<Item = usize> + 'a>

An iterator over the sorted indices of set values in the underlying boolean array good to array with low number of set values.

fn maybe_null_slices_iter<'a>( &'a self, ) -> Box<dyn Iterator<Item = (usize, usize)> + 'a>

An iterator over the sorted disjoint contiguous range set values in the underlying boolean array good for arrays with only long runs of set values.

Provided Methods

fn invert(&self) -> VortexResult<Array>
where Self: Clone,

Return a new inverted version of this array.

True -> False False -> True Null -> Null

fn true_count(&self) -> usize